requirements.txt
torch>=2.2.2
torchvision>=0.17.2
torchaudio>=2.2.2
transformers>=4.35.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
scipy>=1.10.0
numpy>=1.24.0,<2.0
librosa>=0.10.0
soundfile>=0.12.0
pillow>=9.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
wandb>=0.15.0
tensorboard>=2.12.0
tqdm>=4.65.0
einops>=0.6.0
timm>=0.9.0
laion-clap>=1.1.0
openai-whisper>=20230314
evaluate>=0.4.0
pycocotools>=2.0.6